backend/requirements.txt
flask 
requests 
python-dotenv 
flask-cors 
pandas 
sqlalchemy
aiohttp
aiocache